Move observers to its own file
